Update copyright year to 2025 and refactor component imports for consistency
