+++ app/page.tsx
@@ -25,7 +25,7 @@ import { CertificationShowcase } from "@/components/cer";
 import { Hero } from "@/components/hero";
 import LoadingScreen from "@/components/visuals/LoadingScreen";
 import { Background } from "@/components/visuals/motion-background";
-import Technologies from "@/components/tech/Technologies";
+import { Technologies } from "@/components/tech";
 import { ProjectsShowcase } from "@/components/projects";
 import { GitHubShowcase } from "@/components/github";
 import { PageSpeedInsights } from "@/components/pagespeed";
