+++ components/footer/index.ts
@@ -1,0 +1,6 @@
+/**
+ * @author ColdByDefault
+ * @copyright 2025 ColdByDefault. All Rights Reserved.
+ */
+
+export { default as Footer } from "./Footer";
